RAINCATCH-539 - Split client and angular source code
