+++ lib/angular/directives/fileDetailController-spec.js
