+++ lib/angular/directives/fileDetailController.js
