+++ lib/angular/directives/fileDetailDirective.js
