+++ lib/angular/directives/imgDirective.js
