+++ lib/angular/directives/index.js
