+++ lib/angular/file-ng.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var client = require('./file'),
+var client = require('../client/fileClient'),
   _ = require('lodash');
 
 const MODULE_NAME = 'wfm.file.module';
@@ -23,7 +23,14 @@ module.exports = function(userConfig){
     directivesModule
   ]);
   
-  // TODO Remove factory and use mediator
+  wrapClient(mainModule);
+ 
+  require("./portal.js")(mainModule, userConfig);
+  require("./mobile.js")(mainModule, userConfig);
+  return MODULE_NAME;
+};
+
+function wrapClient(mainModule){
   mainModule.factory('fileClient', function($q){
     var fileClient = {};
     _.forOwn(client, function(value, key){
@@ -37,8 +44,4 @@ module.exports = function(userConfig){
     });
     return fileClient;
   });
-  require("./portal.js")(mainModule, userConfig);
-  require("./mobile.js")(mainModule, userConfig);
-  return MODULE_NAME;
-};
-
+}
