+++ lib/angular/templates/file-detail-directive.tpl.html
