+++ lib/angular/templates/file-detail.tpl.html
