+++ lib/angular/templates/file-list.tpl.html
