+++ lib/angular/templates/portal-empty-list.tpl.html
