+++ lib/angular/web-config.js
